refactor upgrade job creation function into new file
